Change the type from int to double for precise calculations.

+++ Chp2/Make_a_difference/Body_mass_index_calculator.cpp
@@ -25,9 +25,9 @@
 using namespace std;
 
 int main() {
-int weightInPounds;
-int heightInInches;
-int BMI;
+double weightInPounds;
+double heightInInches;
+double BMI;
 
 cout << "Enter your weight in pounds: ";
 cin >> weightInPounds;
